+++ views/default/amapnews/group_module.php
@@ -4,6 +4,8 @@
  * @package amapnews
  */
 
+use Amapnews\NewsOptions;
+
 if (!NewsOptions::allowPostOnGroups()) {
     return;
 }