+++ views/default/resources/amapnews/add_existed.php
@@ -4,17 +4,17 @@
  * @package amapnews
  */
 
-elgg_load_css('amapnews_css');
+use Amapnews\NewsOptions;
 
 $connected_guid = elgg_extract('guid', $vars, '');
 $entity = get_entity($connected_guid);
 
-if ($entity && !elgg_instanceof($entity, 'object', 'news')) {
+if ($entity && !$entity instanceof \ElggNews) {
     $vars = amapnews_prepare_form_vars();
     $vars['cguid'] = $connected_guid;
     
     $container = $entity->getContainerEntity();
-    if (NewsOptions::allowPostOnGroups() && elgg_instanceof($container, 'group') && $container->canEdit()) {
+    if (NewsOptions::allowPostOnGroups() && ($container instanceof \ElggGroup) && $container->canEdit()) {
         elgg_set_page_owner_guid($container->getGUID());
         $vars['group_guid'] = $container->getGUID();
     }
@@ -27,5 +27,3 @@ if ($entity && !elgg_instanceof($entity, 'object', 'news')) {
 else {
     echo elgg_echo("amapnews:add:novalidentity");
 }
-
-