+++ views/default/resources/amapnews/custom_list_view.php
@@ -4,6 +4,8 @@
  * @package amapnews
  */
 
+use Amapnews\NewsOptions;
+
 $limit = elgg_extract('limit', $vars, elgg_get_config('default_limit'));
 $featured = get_input('featured', false);