+++ views/default/resources/amapnews/edit.php
@@ -4,16 +4,18 @@
  * @package amapnews
  */
 
+use Amapnews\NewsOptions;
+
 $guid = elgg_extract('guid', $vars, '');
 $entity = get_entity($guid);
 
-if (!elgg_instanceof($entity, 'object', 'news') || !$entity->canEdit()) {
+if (!$entity instanceof \ElggNews || !$entity->canEdit()) {
     elgg_error_response(elgg_echo('amapnews:unknown_amapnews'));
     forward(REFERRER);
 }
 
 $page_owner = elgg_get_page_owner_entity();
-if (elgg_instanceof($page_owner, 'group')) {
+if ($page_owner instanceof \ElggGroup) {
     elgg_push_breadcrumb($page_owner->name, "news/group/$page_owner->guid/all");
 } else {
     elgg_push_breadcrumb(elgg_echo('amapnews:menu'), "news");
@@ -24,12 +26,12 @@ elgg_push_breadcrumb($title);
 
 $form_vars = array('name' => 'amapnews', 'enctype' => 'multipart/form-data');
 $vars = amapnews_prepare_form_vars($entity);
-if (NewsOptions::allowPostOnGroups() && elgg_instanceof($page_owner, 'group') && $page_owner->canEdit()) {
+if (NewsOptions::allowPostOnGroups() && ($page_owner instanceof \ElggGroup) && $page_owner->canEdit()) {
     $vars['group_guid'] = $page_owner->guid;
 }
 $content = elgg_view_form('amapnews/add', $form_vars, $vars);
 
-$body = elgg_view_layout('content', array(
+$body = elgg_view_layout('default', array(
     'filter' => '',
     'content' => $content,
     'title' => $title,