+++ views/default/resources/amapnews/view.php
@@ -16,7 +16,7 @@ if (!$entity) {
 }
 
 $page_owner = elgg_get_page_owner_entity();
-if (elgg_instanceof($page_owner, 'group')) {
+if ($page_owner instanceof \ElggGroup) {
     elgg_push_breadcrumb(elgg_echo('groups'), 'groups');
     elgg_push_breadcrumb($page_owner->name, $page_owner->getURL());
     elgg_push_breadcrumb(elgg_echo('amapnews:menu'), "news/group/$page_owner->guid/all");
@@ -29,7 +29,7 @@ elgg_push_breadcrumb($title);
 
 $content = elgg_view_entity($entity, ['full_view' => true]);
 
-$body = elgg_view_layout('content', array(
+$body = elgg_view_layout('default', array(
     'content' => $content,
     'title' => $title,
     'filter' => '',