+++ views/default/widgets/amapnews_featured/content.php
@@ -4,6 +4,8 @@
  * @package amapnews
  */
 
+use Amapnews\NewsOptions;
+
 // modified to be compatible with widget manager
 $owner = get_entity($vars['entity']->owner_guid);
 
@@ -25,7 +27,7 @@ $options = array(
     ),
 );
 
-if (elgg_instanceof($owner, 'user')) {
+if ($owner instanceof \ElggUser) {
     if (!elgg_in_context('dashboard')) {
         $options['owner_guid'] = $owner->guid;
     }
@@ -71,7 +73,7 @@ if (elgg_instanceof($owner, 'user')) {
         $content .= "</ul>";
     }	
 } 
-elseif (elgg_instanceof($owner, 'group')) {
+elseif ($owner instanceof \ElggGroup) {
     $groupGUID = elgg_get_page_owner_guid();
     $options['container_guid']= $groupGUID;